ticket:1150 - Changing all float fields to double
git-svn-id: file:///home/svn/omero/trunk@3494 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/blitz/test/ome/services/blitz/test/utests/IceMethodInvokerUnitTest.java
@@ -980,10 +980,10 @@ public class IceMethodInvokerUnitTest extends MockObjectTestCase {
     @Test(groups = "ticket:1150")
     public void testRFloatDoesNotGetRounded() throws Exception {
         Objective o = new Objective();
-        o.setLensNA(new Float(1.4));
+        o.setLensNA(new Double(1.4));
         omero.model.Objective o2 = (omero.model.Objective) mapper.handleOutput(
                 Objective.class, o);
-        assertEquals(o.getLensNA().floatValue(), o2.getLensNA().getValue());
+        assertEquals(o.getLensNA().doubleValue(), o2.getLensNA().getValue());
     }
 
     // ~ Helpers
